email/requirements.txt
aiofiles==24.1.0
aiogram==3.21.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiokafka==0.12.0
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
appnope==0.1.4
asgiref==3.9.1
asttokens==3.0.0
async-timeout==5.0.1
asyncio==4.0.0
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
clickhouse-connect==0.8.18
clickhouse-driver==0.2.9
comm==0.2.3
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.16
decorator==5.2.1
dj-database-url==3.0.1
Django==4.2.23
django-cors-headers==4.7.0
django-redis==6.0.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.8.1
dynaconf==3.2.11
et_xmlfile==2.0.0
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.116.1
fonttools==4.59.0
frozenlist==1.7.0
gunicorn==23.0.0
h11==0.16.0
hvac==2.3.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
inflection==0.5.1
ipykernel==6.30.1
ipython==8.18.1
jedi==0.19.2
joblib==1.5.2
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kafka==1.3.5
kafka-python==2.2.15
kiwisolver==1.4.7
knox==0.1.14
loguru==0.7.3
lz4==4.4.4
magic-filter==1.0.12
matplotlib==3.9.4
matplotlib-inline==0.1.7
multidict==6.6.3
nest-asyncio==1.6.0
numpy==2.0.2
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
parso==0.8.4
patsy==1.0.1
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
propcache==0.3.2
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.1
redis==6.4.0
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
scikit-learn==1.7.1
scipy==1.16.1
seaborn==0.13.2
setuptools==80.9.0
shortuuid==1.0.13
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.47.2
statsmodels==0.14.5
telegram==0.0.1
threadpoolctl==3.6.0
tornado==6.5.2
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
validators==0.35.0
wcwidth==0.2.13
wheel==0.45.1
xlsxwriter==3.2.5
yarl==1.20.1
zipp==3.23.0
zstandard==0.24.0
prometheus_client==0.23.1